Restructure API tests to their ownd dir.
